# Optimizing uint operations (architecture independent)