@NikVolf 2016-03-08
```
test u512_add      ... bench:      79,473 ns/iter (+/- 2,289)
test u512_sub      ... bench:         461 ns/iter (+/- 14)
```

why so much difference?