@debris 2016-03-08
Maybe benchmark code was incorrectly inlined?