@NikVolf 2016-03-09
looks good to me, though with next pr we should move fixed hash integers to bigint and return tests