@petevine 2016-03-25
I've updated my benchmark results in https://github.com/ethcore/parity/issues/596#issuecomment-192630065, well done!